Package rename to eosio.spectrum.websocket.api
