--- src/main/java/eos/websocket/api/ActionMessage.java
@@ -1,4 +1,0 @@
-package eos.websocket.api;
-
-public class ActionMessage {
-}
