+++ src/main/java/eosio/spectrum/websocket/api/ActionMessage.java
@@ -1,0 +1,4 @@
+package eosio.spectrum.websocket.api;
+
+public class ActionMessage {
+}
