+++ src/main/java/eosio/spectrum/websocket/api/Events.java
@@ -1,4 +1,4 @@
-package eos.websocket.api;
+package eosio.spectrum.websocket.api;
 
 public enum Events {
     subscribe,
