+++ src/main/java/eosio/spectrum/websocket/api/MessagePublisher.java
@@ -1,4 +1,4 @@
-package eos.websocket.api;
+package eosio.spectrum.websocket.api;
 
 public interface MessagePublisher {
