+++ src/main/java/eosio/spectrum/websocket/api/ServiceMessage.java
@@ -1,11 +1,23 @@
-package eos.websocket.api;
+package eosio.spectrum.websocket.api;
+
+import org.springframework.stereotype.Component;
 
 import java.util.List;
 
-public class SubscribeRequest {
+@Component
+public class ServiceMessage {
 
+    private Events event;
     private String account;
     private List<String> actions;
+
+    public Events getEvent() {
+        return event;
+    }
+
+    public void setEvent(Events event) {
+        this.event = event;
+    }
 
     public String getAccount() {
         return account;
@@ -25,10 +37,10 @@ public class SubscribeRequest {
 
     @Override
     public String toString() {
-        return "SubscribeRequest {" +
-                "account='" + account + '\'' +
+        return "{" +
+                "event=" + event +
+                ", account='" + account + '\'' +
                 ", actions=" + actions +
                 '}';
     }
 }
-
